+++ build_runner/test/asset/cache_test.dart
@@ -5,7 +5,6 @@
 import 'dart:convert';
 
 import 'package:build/build.dart';
-import 'package:build_test/build_test.dart';
 import 'package:test/test.dart';
 
 import 'package:build_runner/src/asset/cache.dart';
@@ -17,8 +16,8 @@ void main() {
   var missingTxt = new AssetId('a', 'missing.txt');
   var fooContent = 'bar';
   var fooUTF8Bytes = UTF8.encode('bar');
-  var assets = <AssetId, DatedValue>{
-    fooTxt: new DatedString('bar'),
+  var assets = <AssetId, dynamic>{
+    fooTxt: 'bar',
   };
   InMemoryRunnerAssetReader delegate;
   CachingAssetReader reader;