+++ build_runner/test/asset_graph/graph_test.dart
@@ -215,8 +215,4 @@ void main() {
 class MockDigestReader extends StubAssetReader with Md5DigestReader {
   @override
   Future<List<int>> readAsBytes(AssetId id) async => UTF8.encode('$id');
-
-  @override
-  Future<DateTime> lastModified(AssetId id) => throw new UnimplementedError(
-      'lastModified not implemented for MockDigestReader');
 }