+++ build_test/lib/src/in_memory_reader.dart
@@ -7,8 +7,6 @@ import 'dart:convert';
 import 'package:build/build.dart';
 import 'package:glob/glob.dart';
 
-import 'in_memory_writer.dart';
-
 /// An [AssetReader] that records which assets have been read to [assetsRead].
 abstract class RecordingAssetReader implements AssetReader {
   Iterable<AssetId> get assetsRead;
@@ -17,7 +15,7 @@ abstract class RecordingAssetReader implements AssetReader {
 /// An implementation of [AssetReader] with primed in-memory assets.
 class InMemoryAssetReader
     implements MultiPackageAssetReader, RecordingAssetReader {
-  final Map<AssetId, DatedValue> assets;
+  final Map<AssetId, List<int>> assets;
   final String rootPackage;
 
   @override
@@ -26,9 +24,19 @@ class InMemoryAssetReader
   /// Create a new asset reader that contains [sourceAssets].
   ///
   /// May optionally define a [rootPackage], which is required for some APIs.
-  InMemoryAssetReader({Map<AssetId, DatedValue> sourceAssets, this.rootPackage})
-      : assets = sourceAssets ?? <AssetId, DatedValue>{},
+  InMemoryAssetReader({Map<AssetId, dynamic> sourceAssets, this.rootPackage})
+      : assets = _convertAssetsToBytes(sourceAssets) ?? <AssetId, List<int>>{},
         assetsRead = new Set<AssetId>();
+
+  static Map<AssetId, List<int>> _convertAssetsToBytes(
+      Map<AssetId, dynamic> original) {
+    var newMap = <AssetId, List<int>>{};
+    original.forEach((id, content) {
+      if (content is String) content = UTF8.encode(content as String);
+      newMap[id] = content as List<int>;
+    });
+    return newMap;
+  }
 
   @override
   Future<bool> canRead(AssetId id) async {
@@ -40,14 +48,14 @@ class InMemoryAssetReader
   Future<List<int>> readAsBytes(AssetId id) async {
     if (!await canRead(id)) throw new AssetNotFoundException(id);
     assetsRead.add(id);
-    return assets[id].bytesValue;
+    return assets[id];
   }
 
   @override
   Future<String> readAsString(AssetId id, {Encoding encoding: UTF8}) async {
     if (!await canRead(id)) throw new AssetNotFoundException(id);
     assetsRead.add(id);
-    return assets[id].stringValue;
+    return UTF8.decode(assets[id]);
   }
 
   @override
@@ -63,10 +71,11 @@ class InMemoryAssetReader
   }
 
   void cacheBytesAsset(AssetId id, List<int> bytes) {
-    assets[id] = new DatedBytes(bytes);
+    assets[id] = bytes;
   }
 
-  void cacheStringAsset(AssetId id, String contents) {
-    assets[id] = new DatedString(contents);
+  void cacheStringAsset(AssetId id, String contents, {Encoding encoding}) {
+    encoding ??= UTF8;
+    assets[id] = encoding.encode(contents);
   }
 }